+++ Changes.md
@@ -1,3 +1,4 @@
+* #936 - highlight goals involved in restriction errors
 * #957 - Allow one to specify non-numeric categories that are not awarded for a particular tournament level
 * #933 - Make buttons on subjective score entry larger
 * #950 - Make Tournament Level a first class object, this will allow parts of the award script to be attached to a tournament level, all tournaments now have a level